Refactor js file names
