+++ js/lookup/get_info.test.js
@@ -1,0 +1,11 @@
+/**
+ * @jest-environment jsdom
+ */
+
+import { options } from "./get_info";
+
+test('options', () => {
+    options
+
+    expect(options.do_body_classes).toBe(false);
+});
